Restore API service in docker-compose and update Dockerfile
The changes made were to uncomment the API service within the docker-compose.yml file and enable its set up, this includes building from the Dockerfile, setting up the environment, ports, and specified dependencies. This was done to allow the API service to be buildable and runnable from a docker-compose command.

Additional changes have been made to the Dockerfile located in the API folder. Previously, it was set up to only restore and build API.csproj. This setup was not adequate as the API project depends on other projects like API.Data, API.Domain, and API.Services. Therefore, these projects have been added to the Dockerfile for copy and restore operations during the docker build.