Refactored IdentityCreationDataDto and added authentication service
Renamed and refactored the `IIdentityCreationDataDto` interface to `IdentityCreationDataDto` class. Making this a class eliminates the need for an interface that is not used polymorphically. Also removed the redundant nullability checks by initializing the properties to empty strings.
Created a new `AuthSessionService` that encapsulates the authentication behaviour and removed such dependency from the `IdentityService` and `AuthSessionsController` classes.
Improved encapsulation separates concerns and makes the code easier to maintain and understand.
Introduced `[EmailAddress]` data annotation in AuthSessionCreateRequest.cs to ensure valid email input.
The commit also includes some renaming, reorganizing and clean up for better readability.