Refactor project structure to provide better encapsulation
In this commit, the structure of the project has been refactored, resulting in the renaming and reorganization of namespaces. The 'API.Data' has been renamed to 'API.Infrastructure', and 'API.Services' to 'API.Application'. This change was made to provide more accurate descriptor of their tasks in the system. 'API.Data' was not just dealing with data, and 'API.Services' was not just providing services.

All the migrations and entities were moved from 'API.Data.Migrations' to 'API.Infrastructure.Migrations' and 'API.Data.Entities' to 'API.Infrastructure.Database.Entities' respectively, to fit the new structure.

All service classes have been moved from 'API.Services' to 'API.Application.Services'. All references to 'API.Data' and 'API.Services' in the solution and code files have been updated to 'API.Infrastructure' and 'API.Application' to maintain consistency.

In addition to maintaining readability, these changes should make the project structure more intuitive and easier to understand for anyone working on the project in the future. It is also expected to reduce coupling and increase cohesion.