Add SignInService for decoupling from SignInManager (#6)
Created a SignInService class as an adapter over the SignInManager/ApplicationUser class from the ASP.NET Identity library and updated all usages, decoupling application code from directly depending on the library.

Reason for change is to make the application code more robust and easy to test by depending on abstractions (ISignInService) and not concrete implementations from third-party library, which might change in future versions. Also, it's a part of an ongoing effort to keep the domain model clean, and keeping all Identity related code isolated in the services layer.