Add FluentValidation for better data integrity
A FluentValidation package was added to the project for improving the data validation mechanisms. FluentValidation provides a simple way of implementing valid data checks and helps improve overall data integrity. It is configured to automatically validate all incoming requests in the API project. Validator classes are dynamically loaded at runtime.