Add paginated fetching of pinned cities
Two new DTOs, PaginatedResultDto and PaginationOptionsDto, were added to implement pagination. PinnedCityService was expanded to include method variants that return paginated results and also to allow retrieving pinned cities for a specific user. Corresponding changes were made in the IPinnedCityService interface. A new endpoint 'Paginated' was added in PinnedCitiesController to handle paginated requests for pinned cities. These changes enhance the readability and maintainability of the code and improve response times for large sets of data.