Add FluentValidation to Swagger documentation
The MicroElements.Swashbuckle.FluentValidation NuGet package was added to the API.csproj in order to provide FluentValidation rules integration to the Swagger API documentation. Changes were also made in the Program.cs file to enable the validation rules in Swagger. This addition ensures that API users are informed of validation rules when using the Swagger UI.