Refactor Weather API services and enhance nullability checks
The structures of ForecastWeatherApiService and CurrentWeatherApiService have been refactored for better code structure and reusability. For shared functionality of URI building, an AbstractWeatherApiService was implemented. These services were also moved to the Infrastructure layer for better organization.

Nullability was added to the GetWeatherDetailsAsync method in both PinnedCityWeatherService and IPinnedCityWeatherService to handle cases where a city by id is not found.

The WeatherApiSettings configuration was updated to include BaseUri and the ApiKey was made into an init property, increasing security. These changes were also reflected in the WeatherAPI params in appsettings.json. These commits work to increase project organization, code reuse, and robustness while increasing configuration security and integrity.